moved HookFlagQuery to own file
